Remove astra-plan plan module
