+++ rust/crates/astra-plan/src/lib.rs
@@ -1,11 +1,9 @@
 //! Astra Plan — goal decomposition engine for breaking complex tasks into subtasks.
 
 pub mod decompose;
-pub mod plan;
 pub mod repository;
 
 pub use decompose::*;
-pub use plan::*;
 pub use repository::{
     CloudPlanRepository, InMemoryPlanRepository, NewStepRun, PlanListFilter, PlanLoadError,
     PlanRepository, PlanStepRun, plan_resume_hint_for_session, plan_resume_prompt_hint,
